Preallocate space when unfolding headers
Summary:
Unfolding headers skips the line-break and indentation characters
so the result will be at worst the same size as the original header.
Since unfolding is implemented as appending characters one by one to
the result bytearray, we can pre-allocate the destination field at
the beginning of the loop.

Reviewers: vkrause

Reviewed By: vkrause

Subscribers: mlaurent, #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D8246