Do not clear parameters when changing the mimetype of a Content-Type header
Summary:
As there is no API to retrieve and restore all parameters, it is otherwise
impossible to change the mimetype while preserving all other parameters.
This is needed for example when fixing wrong mimetype information for
display in KMail (see D8551).

Test Plan:
Unit test of messagelib (which is the primary consumer of this API) still
pass too.

Reviewers: #kde_pim, knauss

Reviewed By: knauss

Subscribers: knauss

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D8568