Deprecate ContentType::contentCategory
This is set in a number of places internally and externally, but it's not
read anywhere.

The getter method and internal storage have been removed, the setter method
is still called externally in a few places that need to be cleaned up first
before we can drop that and the enum here as well.