Add const accessors for KMime::Content headers
The non-const variants create missing headers by default, which makes
those rather error prone to use if you are reading/processing rather
than creating a message. The new const overloads don't do that and
thus are actually usable on a const instance.

This is by far not enough to make the entire Content/Message API
usable on const instances, but it's a first step into that direction.