Add Content::setEncodedBody
This is a replacement API for manually setting the internal decoded state
which is currently done via ContentTransferEncoding::decoded.

The idea is for that state to become fully internal to Content, so it can
be mutable allowing for decoded[Content|Text] to become const, without
having to create a ContentTransferEncoding instance to change that flag
(and thus mutating the actual message).

There's a handful of consumers that need careful adjustment, therefore this
only adds the new API for now. Once everything is ported we can remove
the API in ContentTransferEncoding and change the internal workings as
described above.