Don't search for uuencode begin markers with a QRegularExpression
Also, only create the subject regex when we are actually dealing with
an uuencode part.

This also avoids a conversion of the entire content to a QString in order
to be able to apply the regex.

What is worse, despite this being a parse for a rarely used legacy niche
format, KMime runs this (and all others) unconditionally on any text part,
as those predate proper content type identification headers. So the
performance of this impacts parsing any regular email!