Remove the charset argument in Header::fromUnicodeString()
And same for ContentType::setName().

This is practically everywhere called with UTF-8 as a literal, most
exceptions are in unit test code. Not all call sites use QByteArrayLiteral,
so we also get an additional allocation there.

For the unlikely case this actually needs to be set to something other
than UTF-8, there's still a dedicated setter method for that.

Given the wide-spread use, this has an deprecated inline compat method
for now.