Don't inherit SingleIdent from Ident
We don't rely on that inheritance anywhere, and it allows more efficient
storage of both as well as avoids the extra allocation for the storage
in a list in SingleIdent.

This matters as this is on the hot path for folder loading and message
list population.

There's more that can be done here, such as not disassembling the
identifier into its parts, but that can be done separately without
affecting ABI.