Workaround Outlook server lack of multipart/encrypted support
When sending an email with a multipart/encrypted, outlook will change
the multipart/encrypted MIME type with multipart/mixed. This is likely
happening because multipart/encrypted isn't part of the spec that
Outlook support.

So when encountering a multipart/mixed, also check if none of the
children part is actually encrypted content.