Don't decode content of type text/pgp as text
Content of type text/pgp is either ASCII-armored encrypted text (with or
without CTE) or a base64-encoded binary blob of encrypted text. It must
not be decoded as text because this would append a '\n' to the decoded
content and that breaks decryption (because gpg doesn't like encrypted
data with trailing invalid data). In particular, the erroneously appended
'\n's caused flaky tests in messagelib.