Fix parsing of empty multipart parts
This resulted in QByteArray::mid() called with -1 as second argument,
making the part the entire remaining part of the content without advancing
the parsing position accordingly.

In the extreme case like in the added test file this resulted in
recursively creating a quadratic amount of parts on a tree with the depth
of the amount of parts.

Test provided by oss-fuzz.