Replace custom unique string logic with a UUID
UUIDs ensure all the properties we want here (something we didn't have
when this code was written). Applying Base64 encoding to the UUID data
instead of the usual hex encoding also ensures this doesn't explode in
size compared to the old approach.

As this is also used for the multi-part boundary string we have to make
sure to comply with its grammar, which we do by using Base64 URL encoding,
which unlike the regular one uses only characters allowed here.